reorganised folder structure of frontend/src subfolders
