+++ policy-engine-modules/policy-engine-ui/frontend/src/components/policy-engine-app/PolicyEngineApp.js
@@ -1,5 +1,5 @@
 import React, {useState} from 'react';
-import "../css/Zonesoft.css"
+import "../static/css/Zonesoft.css"
 import AssetTypes from "../asset-types/AssetTypes";
 import Policies from "../policies/Policies";
 import Attributes from "../attributes/Attributes";
@@ -8,6 +8,7 @@ import Attributes from "../attributes/Attributes";
 function PolicyEngineApp() {
 	
 	const [selectedAssetTypes, setSelectedAssetTypes] = useState([]);
+	const [selectedPolicy, setSelectedPolicy] = useState([]);
 	
 	return (
 		<table className="zsft-table" style={{ width: "100%" }}>
@@ -23,8 +24,8 @@ function PolicyEngineApp() {
 					<td colSpan={2}><AssetTypes selectedAssetTypesSetter={setSelectedAssetTypes} /></td>
 				</tr>
 				<tr>
-					<td><Policies selectedAssetTypes={selectedAssetTypes} /></td>
-					<td><Attributes  /></td>
+					<td><Policies selectedAssetTypes={selectedAssetTypes} selectedPolicySetter={setSelectedPolicy}/></td>
+					<td><Attributes selectedPolicy={selectedPolicy} /></td>
 				</tr>
 			</tbody>
 		</table>
