+++ policy-engine-modules/policy-engine-ui/frontend/src/components/policy-engine-app/policies/Policies.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, /*useRef*/ } from 'react';
-import "../css/Zonesoft.css"
+import "../static/css/Zonesoft.css"
 import * as DataService from "./DataService";
 
 function Policies(props) {
@@ -21,6 +21,10 @@ function Policies(props) {
 			console.warn("Policies:useEffect:props.selectedAssetTypes is not set");
 		}
 	}, [props.selectedAssetTypes]);	
+	
+	useEffect(() => {
+		props.selectedPolicySetter(selected);
+	}, [props, selected]);
 	
 	const onAddNewClick = () =>{
 		//to-do
